Removed the `FilterBtn` file entirely. Now we only use the shared `FilterBtnWrapper` component.
I added a prop to the `Modal` component because it was far too wide for 
displaying just the filter options.
